Suppress conversion warnings

+++ dataobj/koord3d.h
@@ -117,19 +117,19 @@ static inline koord3d operator - (const koord3d& a, const koord& b)
 
 static inline uint32 koord_distance(koord3d a, koord b)
 {
-	return abs(a.x - b.x) + abs(a.y - b.y);
+	return (uint32)(abs(a.x - b.x)) + (uint32)(abs(a.y - b.y));
 }
 
 
 static inline uint32 koord_distance(koord a, koord3d b)
 {
-	return abs(a.x - b.x) + abs(a.y - b.y);
+	return (uint32)(abs(a.x - b.x)) + (uint32)(abs(a.y - b.y));
 }
 
 
 static inline uint32 koord_distance(koord3d a, koord3d b)
 {
-	return abs(a.x - b.x) + abs(a.y - b.y);
+	return (uint32)(abs(a.x - b.x)) + (uint32)(abs(a.y - b.y));
 }
 
 /**
